Split entries tests into focused modules
